Refactoring in deploymodel.tests project.
Clean-up of some Xpect-related structures.
