+++ tests/org.franca.deploymodel.dsl.tests/src/org/franca/deploymodel/dsl/tests/xpect/ui/ContentAssistUITest.fdepl.xt
@@ -1,14 +1,14 @@
 /* 
 XPECT_SETUP org.franca.deploymodel.dsl.tests.xpect.ui.FrancaDeploymodelXpectUITests
 Workspace{
-	JavaProject "libraryproject"{
+	org.xpect.xtext.lib.setup.workspace.JavaProject "libraryproject" {
 		nature = "org.eclipse.xtext.ui.shared.xtextNature"	
 		builder = "org.eclipse.xtext.ui.shared.xtextBuilder"
 		SrcFolder "model" {
 			File "d.fdepl" {}
 		}
 	}
-	org.franca.deploymodel.dsl.tests.xpect.JavaProject "sample"{
+	org.franca.deploymodel.dsl.tests.xpect.util.JavaProjectWithReferences "sample" {
 		projectReference = "libraryproject"
 		nature = "org.eclipse.xtext.ui.shared.xtextNature"	
 		builder = "org.eclipse.xtext.ui.shared.xtextBuilder"
