+++ tests/org.franca.deploymodel.dsl.tests/src/org/franca/deploymodel/dsl/tests/xpect/ui/FrancaDeploymodelXpectUITests.java
@@ -41,7 +41,7 @@ import com.google.inject.Injector;
 import com.google.inject.Provider;
 
 @SuppressWarnings("restriction")
-@XpectSuiteClasses({ FrancaDeploymodelXpectTests.class})
+@XpectSuiteClasses({ FrancaDeploymodelXpectTests.class })
 @RunWith(XpectRunner.class)
 @XpectTestFiles(baseDir=".",  fileExtensions = "xt")
 @XpectSetup({ XtextWorkspaceSetup.class })
