+++ tests/org.franca.deploymodel.dsl.tests/src/org/franca/deploymodel/dsl/tests/xpect/util/JavaProjectWithReferences.java
@@ -1,4 +1,4 @@
-package org.franca.deploymodel.dsl.tests.xpect;
+package org.franca.deploymodel.dsl.tests.xpect.util;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -7,34 +7,38 @@ import java.util.List;
 import org.eclipse.core.resources.IProject;
 import org.eclipse.core.resources.IWorkspaceRoot;
 import org.eclipse.core.runtime.CoreException;
-import org.eclipse.core.runtime.Path;
 import org.eclipse.jdt.core.IJavaProject;
 import org.eclipse.jdt.core.JavaCore;
 import org.eclipse.xtext.junit4.ui.util.JavaProjectSetupUtil;
 import org.xpect.xtext.lib.setup.FileSetupContext;
 import org.xpect.xtext.lib.setup.workspace.Workspace;
-/** Extends Xpect's {@link JavaProject} in order to provide projectReferences as additional feature.<br/>
+
+/**
+ * Extends Xpect's {@link JavaProject} in order to provide projectReferences as additional feature.<br/>
+ * 
  * A setup might look like
  * <pre>
-	XPECT_SETUP org.franca.deploymodel.dsl.tests.xpect.ui.FrancaDeploymodelXpectUITests
+	XPE CT_SETUP org.franca.deploymodel.dsl.tests.xpect.ui.FrancaDeploymodelXpectUITests
 	Workspace{
-		JavaProject "libraryproject"{ ...	}
-		org.franca.deploymodel.dsl.tests.xpect.JavaProject "sample"{ 
+		JavaProject "libraryproject" { ... }
+		JavaProjectWithReferences "sample" {
 			projectReference = "libraryproject" 
 			...
 		}
 	} 
-	END_SETUP
+	END _SETUP
  </pre>
-  */ 
-public class JavaProject extends org.xpect.xtext.lib.setup.workspace.JavaProject {
+ */ 
+@SuppressWarnings("restriction")
+public class JavaProjectWithReferences extends org.xpect.xtext.lib.setup.workspace.JavaProject {
+
 	protected List<String> projectReferences = new ArrayList<String>();
 
-	public JavaProject(String name) {
+	public JavaProjectWithReferences(String name) {
 		super(name);
 	}
 
-	public JavaProject() {
+	public JavaProjectWithReferences() {
 		super();
 	}
